Add Generating summary section Update primary instruction and Add New prompt instructions
