+++ pyproject.toml
@@ -91,6 +91,7 @@ colorlog = "*"
 mapbox-tilesets = "*"
 ipython = "*"
 tiktoken = "*"
+openai = "*"
 
 [tool.poetry.dev-dependencies]
 pytest-profiling = "*"
